Move tests to correct file
